add backbuffer to prevent tearing

+++ src/gameboy.rs
@@ -184,7 +184,7 @@ impl GameBoy {
         }
     }
 
-    pub fn get_frame_buffer(&self) -> &[u32] {
+    pub fn get_frame_buffer(&self) -> &[u8] {
         self.ppu.get_frame_buffer()
     }
